logging; helper refactor part1; scheduler enable_structured=True; rename example folder (#28)
