+++ nncf/experimental/torch/sparsity/movement/algo.py
@@ -21,10 +21,10 @@ from nncf.api.compression import CompressionStage
 from nncf.common.accuracy_aware_training.training_loop import ADAPTIVE_COMPRESSION_CONTROLLERS
 from nncf.common.graph import NNCFNode
 from nncf.common.graph.transformations.commands import TargetType
+from nncf.common.logging import nncf_logger
 from nncf.common.sparsity.statistics import MovementSparsityStatistics
 from nncf.common.statistics import NNCFStatistics
 from nncf.common.utils.helpers import matches_any
-from nncf.common.utils.logger import logger
 from nncf.config.extractors import extract_algo_specific_config
 from nncf.experimental.torch.sparsity.movement.layers import MovementSparsifier
 from nncf.experimental.torch.sparsity.movement.layers import SparseConfig
@@ -85,10 +85,10 @@ class MovementSparsityBuilder(BaseSparsityAlgoBuilder):
             node_name = module_node.node_name
 
             if not self._should_consider_scope(node_name):
-                logger.info('Ignored adding Weight Sparsifier in scope: %s', node_name)
+                nncf_logger.info('Ignored adding Weight Sparsifier in scope: %s', node_name)
                 continue
 
-            logger.debug('Adding Weight Sparsifier in scope: %s', node_name)
+            nncf_logger.debug('Adding Weight Sparsifier in scope: %s', node_name)
             compression_lr_multiplier = self.config.get_redefinable_global_param_value_for_algo(
                 'compression_lr_multiplier', self.name)
             sparsifying_operation = self.create_weight_sparsifying_operation(module_node, compression_lr_multiplier)
